requirements.txt
asgiref==3.2.5
blis==0.4.1
catalogue==1.0.0
certifi==2019.11.28
chardet==3.0.4
cymem==2.0.3
Django==3.0.4
djangorestframework==3.11.0
idna==2.9
importlib-metadata==1.5.0
murmurhash==1.0.2
numpy==1.18.2
plac==1.1.3
preshed==3.0.2
pytz==2019.3
requests==2.23.0
spacy==2.2.4
spacy-lookups-data==0.2.0
sqlparse==0.3.1
srsly==1.0.2
thinc==7.4.0
tqdm==4.43.0
urllib3==1.25.8
wasabi==0.6.0
zipp==3.1.0
psycopg2==2.8.5
